CalabiYau-tools/requirements.txt
torch>=2.2,<3
ultralytics>=8.0,<9
opencv-python>=4.8,<5
numpy>=1.24,<3
pyserial>=3.5,<4
